chore: refine code
